Place backend server operations into a separate file, make behavior opt-out
